requirements.txt
folium==0.14.0
geopy==2.4.0
numpy==1.24.4
pandas==2.0.3
Pillow==9.5.0
# Pillow==10.0.0
pytest==7.4.0
Requests==2.31.0
streamlit==1.25.0
streamlit_folium==0.13.0
